(maint) Update Rakefile to make rspec optional
Currently rspec is a hard dependency to using the Rakefile for puppet.
This means that users on platforms without native rspec packages can't
use the Rakefile to do other things, like build packages.  This commit
moves the require on rspec and rspec/core/rake_task into a begin/rescue
block and wraps the call to Rspec::Core::RakeTask in a branching
conditional so it is only invoked if Rspec is loaded and available.