Add Facter::Util::FileRead.read helper method
Without this patch, there isn't a robust and easy way to simply read the
contents of a file from within a fact's setcode block.  This is a
problem because it encourages the use of
Facter::Util::Resolution.exec('cat ...') which is a bit heavy for this
purpose.

This patch addresses the problem by adding a new utility method and
module intended to simply read and return the contents of a file.  If
the file is inaccessible or does not exist then the method logs a debug
message and returns nil.