Added Expected behavior for operatingsystemrelease on Ubuntu
Without this patch, there are no examples of the expected behavior of
the operatingsystemrelease fact for Ubuntu.  This is a problem because
it is unclear if the fact's value should contain the patch version in
addition to the major and minor version.

This patch fixes the problem by adding an RSpec example that clearly
specifies the operatingsystemrelease fact should only contain the major
and minor version and not the patch version.  If this behavior changes
in the future this example will fail to pass.

See https://github.com/puppetlabs/facter/pull/278 for the decision.