WIP Improve Ubuntu operatingsystemrelease resolution
Without this patch the operatingsystemrelease fact on Ubuntu requires
the full 3 versions.  This is a problem because the initial release may
contain only on the major and minor version numbers in the issue file.

This patch fixes the problem by switching the regular expression to
treat the third patch version identifier as optional.

WIP pending examples of the new behavior.  Here's the issue file to test
against:  "Ubuntu 10.04.4 LTS \\n \\l\n\n"