Make json an optional bundler dependency
Without this patch I cannot satisfy all of the dependencies on AIX
because I do not have a compiler which is required to install compiled
libraries.

This patch works around the issue with Facter on AIX by making the JSON
dependency an optional dependency.  I believe this to be acceptable
because the only place JSON appears to be used in Facter is when
outputting JSON on the command line.