Fix newline in operatingsystemrelease fact
Without this patch applied there is a trailing newline in the value of
the `operatingsystemrelease` Facter fact on Debian systems.  The root
case of this problem is the change made in df71d88 which replease
useless use of cat with File.read methods.

This patch addresses the problem by stripping off all trailing
whitespace from the affected operatingsystemrelease resolvers.