Add gce facts from metadata server
Without this patch Facter does not populate the fact values with
information available in the Google Compute Engine metadata server.
This is a problem because end users will need to configure their
compute instances based on information provided by the metadata service.

This patch addresses the problem by reading the values from the metadata
server only if the DMI data contains the google compute identifier as
described in https://developers.google.com/compute/docs/instances#dmi
This behavior is similar to the AWS behavior and is improved over the
AWS behavior because we do not need to rely on a timeout to determine
the availability of the metadata service.  The DMI data alone is enough
to inform Facter of the metadata service presence.